+++ probe-rs/src/architecture/arm/sequences/nrf91.rs
@@ -1,29 +1,28 @@
-//! Sequences for the nRF53.
+//! Sequences for the nRF91.
 
 use std::sync::Arc;
 
 use super::{nrf::Nrf, ArmDebugSequence};
-use crate::architecture::arm::ap::CSW;
 use crate::architecture::arm::{
     communication_interface::Initialized, ApAddress, ArmCommunicationInterface, DapAccess,
 };
 use crate::Memory;
 
-/// The sequence handle for the nRF5340.
-pub struct Nrf5340(());
+/// The sequence handle for the nRF9160.
+pub struct Nrf9160(());
 
-impl Nrf5340 {
-    /// Create a new sequence handle for the nRF5340.
+impl Nrf9160 {
+    /// Create a new sequence handle for the nRF9160.
     pub fn create() -> Arc<dyn ArmDebugSequence> {
         Arc::new(Self(()))
     }
 }
 
-impl Nrf for Nrf5340 {
+impl Nrf for Nrf9160 {
     fn core_aps(&self, interface: &mut Memory) -> Vec<(ApAddress, ApAddress)> {
         let ap_address = interface.get_ap();
 
-        let core_aps = [(0, 2), (1, 3)];
+        let core_aps = [(0, 4)];
 
         core_aps
             .into_iter()
@@ -45,16 +44,14 @@ impl Nrf for Nrf5340 {
     fn is_core_unlocked(
         &self,
         arm_interface: &mut ArmCommunicationInterface<Initialized>,
-        ahb_ap_address: ApAddress,
-        _ctrl_ap_address: ApAddress,
+        _ahb_ap_address: ApAddress,
+        ctrl_ap_address: ApAddress,
     ) -> Result<bool, crate::Error> {
-        let csw: CSW = arm_interface
-            .read_raw_ap_register(ahb_ap_address, 0x00)?
-            .into();
-        Ok(csw.DeviceEn != 0)
+        let approtect_status = arm_interface.read_raw_ap_register(ctrl_ap_address, 0x00C)?;
+        Ok(approtect_status != 0)
     }
 
     fn has_network_core(&self) -> bool {
-        true
+        false
     }
 }